authentication is now added in expense application

+++ controllers/expense.js
@@ -9,7 +9,9 @@ exports.postexpense = async(req, res) => {
         return res.status(400).json({success: false, message: "parameters missing"})
     }
 
-    const data = await expenses.create({expense, description, category});
+    // const userId = req.user;
+    // const data = await expenses.create({expense, description, category, userId: req.user.id});
+    const data = await req.user.createExpense({expense, description, category});
     res.status(200).json({success: true, message: "expenses successfully added", allData: data});
 
 }
@@ -18,7 +20,11 @@ exports.getexpense = async(req, res, next) => {
 
     try{
 
-    const data = await expenses.findAll();
+    // const users = req.user;
+    // console.log("The user is "+ users);
+    const data = await req.user.getExpenses();
+    // const data = await expenses.findAll({where: {userId: req.user.id}});
+    // const data = await expenses.findAll();
     console.log("hello world");
     // console.log(data);
     return res.status(200).json({success: true, data});
@@ -42,7 +48,10 @@ exports.detetePost = async (req, res, next) => {
             res.status(400).json({success: false})
         }
 
-        await expenses.destroy({where: {id: userId}})
+        const rows = await expenses.destroy({where: {id: userId, userId: req.user.id}})
+        if(rows === 0){
+            return res.status(404).json({success: false, message: "Expense is not belong to the user"});
+        }
         return res.status(200).json({success: true, message: "deleted successfully"});
         
     }
